requirements.txt
<<<<<<< HEAD
fastapi
uvicorn
=======
﻿aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pytz==2024.1
yarl==1.9.4
python-dotenv
>>>>>>> 55610b225b09959c9733238c99358c01d3171652
